+++ main/Spawn.cpp
@@ -4,13 +4,13 @@
 #include "Logger.hpp"
 #include "Sentinel.hpp"
 
-Spawn::Spawn(const ObjectType spawnType, const Sentinel& prototype, const unsigned int size,
+Spawn::Spawn(const ObjectType spawnType, const Sentinel& prototype, const unsigned short size,
 	const Color24 color) :
 	WorldObject(spawnType, color)
 {
 	// prototypes are guaranteed to be squares, so \Delta x = \Delta y
-	const unsigned int prototypeSize = prototype.GetBounds().max.x - prototype.GetBounds().min.x;
-	const int sizeDiff = prototypeSize - size;
+	const unsigned short prototypeSize = prototype.GetBounds().max.x - prototype.GetBounds().min.x;
+	const short sizeDiff = prototypeSize - size;
 	if(sizeDiff < 0)
 	{
 		Logger::Debug(boost::format("Spawn of type %1% was passed a sentinel smaller than the spawn size!")
